Replace diacriticless-Module with optimized internal alternative (#842)
As the npm-Module "diacriticless" is highly inefficient when replacing
dicritics (it iterates thousands of times multiple arrays) and the
author does not merge the proposed pull-request with a very optimized
routine of this, this optimization is now implemented locally in
vue-good-table. As https://github.com/xaksis/vue-good-table/issues/545
states, there is a performanceproblem because of the
diacriticless-implementation. So this could help fixing this issue as
well.